feat: remove workarounds for Chrome v113 and below
Chrome up until v113 had a bug which caused it to emit an empty blob instead of an error when
recording a MediaStream with a video track as audio. It also dispatched a dataavailable and a stop
event before dispatching an error event. The workarounds for those bugs get removed now.

BREAKING CHANGE: This will remove support for Chrome v113.